% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_options.R
\name{ql_disable_db}
\alias{ql_disable_db}
\title{Disable caching for the current session}
\usage{
ql_disable_db()
}
\value{
Nothing, used for its side effects.
}
\description{
Disable caching for the current session
}
\examples{
ql_disable_db()
}
\seealso{
Other database: 
\code{\link{ql_enable_db}()},
\code{\link{ql_set_db_options}()}
}
\concept{database}
