% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_set_api_keys.R
\name{ql_get_api_keys}
\alias{ql_get_api_keys}
\title{Retrieve previously set Ollama API keys typically used for cloud models}
\usage{
ql_get_api_keys(api_keys = NULL)
}
\arguments{
\item{api_keys}{Valid Ollama API keys, typically used for cloud models.}
}
\value{
API keys if previously set with \code{ql_set_api_keys()}, an empty string
if not set, or the same keys provided as input to this function.
}
\description{
Retrieve previously set Ollama API keys typically used for cloud models
}
\examples{
ql_set_api_keys("<your_api_keys_here>")
ql_get_api_keys()
}
