% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_options.R
\name{ql_get_db_options}
\alias{ql_get_db_options}
\title{Retrieve}
\usage{
ql_get_db_options(options = c("db", "db_type", "db_folder", "db_name"))
}
\arguments{
\item{options}{Available options that}
}
\value{
A list with the selected options.
}
\description{
Retrieve
}
\examples{
ql_get_db_options()

## Retrieve only selected option
ql_get_db_options("db_type")
}
