% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_options.R
\name{ql_get_options}
\alias{ql_get_options}
\title{Get options}
\usage{
ql_get_options(
  options = c("model", "host", "system", "temperature", "seed"),
  system = NULL,
  host = NULL,
  model = NULL,
  temperature = NULL,
  seed = NULL
)
}
\arguments{
\item{options}{A character vector with all available options.}
}
\value{
A list with all available options (or those selected with
\code{options})
}
\description{
Get options
}
\examples{


ql_set_options(
  model = "llama3.2",
  host = "http://localhost:11434",
  system = "You are a helpful assistant.",
  temperature = 0,
  seed = 42
)

ql_get_options()
}
