% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_hash.R
\name{ql_hash}
\alias{ql_hash}
\title{Hash all inputs relevant to the call to the LLM and create a hash to be used
for caching.}
\usage{
ql_hash(prompt_df)
}
\arguments{
\item{prompt_df}{A data frame with all inputs passed to the LLM, typically
created with \code{\link[=ql_prompt]{ql_prompt()}}.}
}
\value{
A tibble, such as those returned by \code{\link[=ql_prompt]{ql_prompt()}}, but always
including a hash column.
}
\description{
Mostly used internally.
}
\examples{
ql_prompt("a haiku", hash = FALSE) |> ql_hash()
}
