% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_read_images.R
\name{ql_read_images}
\alias{ql_read_images}
\title{Read image in order to pass it to multimodal models}
\usage{
ql_read_images(path)
}
\arguments{
\item{path}{Path to image file.}
}
\value{
A list object of character vectors of base 64 encoded images.
}
\description{
Read image in order to pass it to multimodal models
}
\examples{
if (interactive) {
  library("quackingllama")

  img_path <- fs::file_temp(ext = "png")

  download.file(
    url = "https://ollama.com/public/ollama.png",
    destfile = img_path
  )

  resp_df <- ql_prompt(
    prompt = "what is this?",
    images = img_path,
    model = "llama3.2-vision"
  ) |>
    ql_generate()

  resp_df

  resp_df$response
}
}
