% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_request.R
\name{ql_request}
\alias{ql_request}
\title{Create \code{httr2} request for both generate and chat endpoints}
\usage{
ql_request(
  prompt_df,
  endpoint = "generate",
  host = NULL,
  messages = NULL,
  keep_alive = NULL,
  timeout = NULL
)
}
\arguments{
\item{endpoint}{Defaults to "generate". Must be either "generate" or "chat".}

\item{host}{The address where the Ollama API can be reached, e.g.
\verb{http://localhost:11434} for locally deployed Ollama.}

\item{messages}{Defaults to NULL. If given, in line with official Ollama
documentation: "the messages of the chat, this can be used to keep a chat
memory".}

\item{keep_alive}{Defaults to "5m". Controls how long the model will stay
loaded into memory following the request.}

\item{timeout}{If not set with \code{\link[=ql_set_options]{ql_set_options()}}, defaults to 300 seconds (5
minutes).}
}
\value{
A \code{httr2} request object.
}
\description{
Create \code{httr2} request for both generate and chat endpoints
}
\examples{
ql_prompt(prompt = "a haiku")

ql_prompt(prompt = "a haiku") |>
  ql_request() |>
  httr2::req_dry_run()
}
