% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_set_api_keys.R
\name{ql_set_api_keys}
\alias{ql_set_api_keys}
\title{Set Ollama API keys typically used for cloud models}
\usage{
ql_set_api_keys(api_keys)
}
\arguments{
\item{api_keys}{Valid Ollama API keys, typically used for cloud models.}
}
\value{
Invisibly returns API keys, mainly used for its side effects.
}
\description{
Set Ollama API keys typically used for cloud models
}
\examples{
ql_set_api_keys("<your_api_keys_here>")
}
