% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_options.R
\name{ql_set_db_options}
\alias{ql_set_db_options}
\title{Set options for the local database}
\usage{
ql_set_db_options(
  db_name = "quackingllama",
  db_type = "DuckDB",
  db_folder = "."
)
}
\arguments{
\item{db_name}{Defaults to \code{quackingllama}. Name given to the local database
file. Useful for differentiating among different approaches or projects
when storing multiple database files in the same folder.}

\item{db_type}{Defaults to \code{DuckDB}.}

\item{db_folder}{Defaults to \code{.}, i.e., to the current working directory.}
}
\value{
Nothing, used for its side effects..
}
\description{
Set options for the local database
}
\examples{
ql_set_db_options(db_name = "testing_ground")
}
\seealso{
Other database: 
\code{\link{ql_disable_db}()},
\code{\link{ql_enable_db}()}
}
\concept{database}
