% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_options.R
\name{ql_set_options}
\alias{ql_set_options}
\title{Set basic options for the current session.}
\usage{
ql_set_options(
  model = NULL,
  host = NULL,
  system = NULL,
  temperature = NULL,
  seed = NULL
)
}
\arguments{
\item{seed}{}
}
\description{
Set basic options for the current session.
}
