% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_translate.R
\name{ql_translate}
\alias{ql_translate}
\title{Translate strings relying on dedicated models}
\usage{
ql_translate(
  text,
  source_language = NULL,
  source_language_code = NULL,
  target_language = NULL,
  target_language_code = NULL,
  translation_model = "translategemma:4b",
  temperature = NULL,
  seed = NULL,
  host = NULL,
  hash = TRUE,
  only_cached = FALSE,
  keep_alive = NULL,
  timeout = NULL,
  error = c("fail", "warn")
)
}
\arguments{
\item{text}{Text to be translated.}

\item{source_language}{Defaults to \code{NULL}. If not given, detected
automatically. If \code{source_language_code} given, \code{source_language} is
ignored. Source language as a full word, e.g. "English". For available
languages with the default \code{translategemma} model, see the list of
available languages in the
\href{https://ollama.com/library/translategemma}{documentation on Ollama's
website}.}

\item{source_language_code}{Defaults to \code{NULL}. If not given, detected
automatically. If \code{source_language_code} given, \code{source_language} is
ignored. Source language as a langauge code, e.g. "en", or "en-GB". For
available languages with the default \code{translategemma} model, see the list of
available languages in the
\href{https://ollama.com/library/translategemma}{documentation on Ollama's
website}.}

\item{target_language}{Defaults to \code{NULL}. If \code{target_language_code} given,
\code{target_language} is ignored. Target language as a full word, e.g. "English".
For available languages with the default \code{translategemma} model, see the
list of available languages in the
\href{https://ollama.com/library/translategemma}{documentation on Ollama's
website}.}

\item{target_language_code}{Defaults to \code{en}. If not given, detected
from  \code{target_language}. Either \code{target_language} or \code{target_language_code}
must be given. Target language as a langauge code, e.g. "en", or "en-GB".
For available languages with the default \code{translategemma} model, see the
list of available languages in the
\href{https://ollama.com/library/translategemma}{documentation on Ollama's
website}.}

\item{translation_model}{Defaults to \verb{translategemma:4b} (12b and 27 are also
available). May work also with other models not specifically targeting
translation.}

\item{temperature}{Numeric value comprised between 0 and 1 passed to the
model. When set to 0 and with the same seed, the response to the same
prompt is always exactly the same. When closer to one, the response is more
variable and creative. Use 0 for consistent responses. Setting this to 0.7
is a common choice for creative or interactive tasks.}

\item{seed}{An integer. When temperature is set to 0 and the seed is
constant, the model consistently returns the same response to the same
prompt.}

\item{host}{The address where the Ollama API can be reached, e.g.
\verb{http://localhost:11434} for locally deployed Ollama.}

\item{hash}{Defaults to \code{TRUE}. If \code{TRUE}, adds a column with the hash of all
other components of the prompt. Used internally for caching. Can be added
separately with \code{\link[=ql_hash]{ql_hash()}}.}

\item{only_cached}{Defaults to \code{FALSE}. If \code{TRUE}, only cached responses are
returned.}

\item{keep_alive}{Defaults to "5m". Controls how long the model will stay
loaded into memory following the request.}

\item{timeout}{If not set with \code{\link[=ql_set_options]{ql_set_options()}}, defaults to 300 seconds (5
minutes).}

\item{error}{Defines how errors should be handled, defaults to "fail", i.e.
if an error emerges while querying the LLM, the function stops. If set to
"warn", it sets the response to \code{NA_character_} and stores it in database.
This can be useful e.g. for proceed if the prompts include a request that
routinely time outs without giving a response. This does not imply that the
model would never give a respones, e.g. re-running the same query with
longer time out may work.}
}
\value{
A data frame such as those returned by \code{\link[=ql_generate]{ql_generate()}} with the
translation of the given input text in the response column.
}
\description{
Translate strings relying on dedicated models
}
\examples{
\dontrun{
translation <- ql_translate(
  text = "A new collection of open translation models built on Gemma 3, helping people communicate across 55 languages.",
  target_language = "french"
)
translation$response
}
}
