% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qt_generate.R
\name{qt_generate}
\alias{qt_generate}
\title{Generate a response and return the result in a data frame}
\usage{
qt_generate(
  prompt,
  system = "You are a helpful assistant.",
  host = "http://localhost:11434",
  model = "llama3.2",
  endpoint = "api/generate",
  temperature = 0,
  seed = NULL
)
}
\arguments{
\item{seed}{}
}
\description{
Generate a response and return the result in a data frame
}
