% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quackingllama-package.R
\docType{package}
\name{quackingllama-package}
\alias{quackingllama}
\alias{quackingllama-package}
\title{quackingllama: Process text with Ollama, retrieve structured results, cache them locally in DuckDB.}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Process text with Ollama, store results in DuckDB.
}
\author{
\strong{Maintainer}: Giorgio Comai \email{g@giorgiocomai.eu} (\href{https://orcid.org/0000-0002-0515-9542}{ORCID}) [copyright holder]

}
\keyword{internal}
