% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_create_schema_enum.R
\name{ql_create_schema_enum}
\alias{ql_create_schema_enum}
\title{Create a schema with one or more questions, and a single set of answers}
\usage{
ql_create_schema_enum(questions, answers)
}
\arguments{
\item{questions}{A character vector, of length 1 or more, with questions to be asked, usually about a text given as prompt.}

\item{answers}{A character vector: the response to each question should be one selected among the given options.}
}
\value{
A list object, that can be passed to the \code{format} argument of \link{ql_prompt}
}
\description{
Create a schema with one or more questions, and a single set of answers
}
\examples{


library("quackingllama")
schema_l <- ql_create_schema_enum(
  questions = c(
    "Does this story include humans",
    "Does this story include animals",
    "Is this story about a woman",
    "Does this story have a happy ending"
  ),
  answers = c("Yes", "No", "Maybe", "Cannot answer")
)


schema_l

\dontrun{
responses_df <- ql_prompt(
  prompt = "this it the story of a duck that escapes",
  format = schema_l,
) |>
  ql_generate() |>
  dplyr::pull("response") |>
  yyjsonr::read_json_str() |>
  tibble::as_tibble()

  responses_df



responses_df <- ql_prompt(
  prompt = "this it the story of a duck that escapes from a lady",
  format = schema_l
) |>
  ql_generate() |>
  dplyr::pull("response") |>
  yyjsonr::read_json_str() |>
  tibble::as_tibble()

  responses_df
}
}
