% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qt_generate.R
\name{ql_generate}
\alias{ql_generate}
\title{Generate a response and return the result in a data frame}
\usage{
ql_generate(
  prompt,
  system = "You are a helpful assistant.",
  host = "http://localhost:11434",
  model = "llama3.2",
  endpoint = "api/generate",
  temperature = 0,
  seed = NULL
)
}
\arguments{
\item{system}{Defaults to the generic: "You are a helpful assistant."}

\item{host}{Defaults to "http://localhost:11434", where locally deployed Ollama usually responds.}

\item{seed}{}
}
\description{
Generate a response and return the result in a data frame
}
