% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_generate.R
\name{ql_generate}
\alias{ql_generate}
\title{Generate a response and return the result in a data frame}
\usage{
ql_generate(
  prompt,
  system = NULL,
  host = NULL,
  model = NULL,
  temperature = NULL,
  seed = NULL,
  endpoint = "api/generate"
)
}
\arguments{
\item{prompt}{A prompt for the LLM.}

\item{seed}{}
}
\value{
A data frame, including a response column, as well as other information returned by the model.
}
\description{
Generate a response and return the result in a data frame
}
\examples{

ql_generate("a haiku", seed = 1)
}
