% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_generate.R
\name{ql_generate}
\alias{ql_generate}
\title{Generate a response and return the result in a data frame}
\usage{
ql_generate(
  prompt_df,
  only_cached = FALSE,
  host = NULL,
  message = NULL,
  timeout = NULL
)
}
\arguments{
\item{prompt_df}{A data frame with all inputs passed to the LLM, typically
created with {ql_prompt()}.}

\item{only_cached}{Defaults to FALSE. If TRUE, only cached responses are
returned.}

\item{host}{The address where the Ollama API can be reached, e.g.
\verb{http://localhost:11434} for locally deployed Ollama.}

\item{timeout}{If not set with {ql_set_options()}, defaults to 300 seconds (5
minutes).}
}
\value{
A data frame, including a response column, as well as other
information returned by the model.
}
\description{
Generate a response and return the result in a data frame
}
\examples{
ql_prompt("a haiku") |>
  ql_generate()
}
