% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_generate.R
\name{ql_generate}
\alias{ql_generate}
\title{Generate a response and return the result in a data frame}
\usage{
ql_generate(
  prompt_df,
  only_cached = FALSE,
  host = NULL,
  message = NULL,
  timeout = NULL,
  error = c("fail", "warn")
)
}
\arguments{
\item{prompt_df}{A data frame with all inputs passed to the LLM, typically
created with \code{\link[=ql_prompt]{ql_prompt()}}.}

\item{only_cached}{Defaults to FALSE. If TRUE, only cached responses are
returned.}

\item{host}{The address where the Ollama API can be reached, e.g.
\verb{http://localhost:11434} for locally deployed Ollama.}

\item{timeout}{If not set with \code{\link[=ql_set_options]{ql_set_options()}}, defaults to 300 seconds (5
minutes).}

\item{error}{Defines how errors should be handled, defaults to "fail", i.e.
if an error emerges while querying the LLM, the function stops. If set to
"warn", it sets the response to \code{NA_character_} and stores it in database.
This can be useful e.g. for proceed if the prompts include a request that
routinely time outs without giving a response. This does not imply that the
model would never give a respones, e.g. re-running the same query with
longer time out may work.}
}
\value{
A data frame, including a response column, as well as other
information returned by the model.
}
\description{
Generate a response and return the result in a data frame
}
\examples{
\dontrun{
ql_prompt("a haiku") |>
  ql_generate()
  }
}
