% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_options.R
\name{ql_get_options}
\alias{ql_get_options}
\title{Get options}
\usage{
ql_get_options(
  options = c("system", "model", "think", "host", "temperature", "seed", "keep_alive",
    "timeout"),
  system = NULL,
  model = NULL,
  think = NULL,
  host = NULL,
  temperature = NULL,
  seed = NULL,
  keep_alive = NULL,
  timeout = NULL
)
}
\arguments{
\item{options}{A character vector used to filter which options should
effectively be returned. Defaults to all available.}
}
\value{
A list with all available options (or those selected with
\code{options})
}
\description{
Get options
}
\examples{


ql_set_options(
  model = "llama3.2",
  host = "http://localhost:11434",
  system = "You are a helpful assistant.",
  temperature = 0,
  seed = 42,
  keep_alive = "5m"
)

ql_get_options()
}
