% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ql_options.R
\name{ql_get_options}
\alias{ql_get_options}
\title{Get options}
\usage{
ql_get_options(options = c("model", "host", "system", "temperature", "seed"))
}
\arguments{
\item{options}{A character vector with all available options}
}
\value{
A list with all available options (or those selections with
\code{options})
}
\description{
Get options
}
\examples{
ql_get_options()
}
